Upd table relationships

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -6,13 +6,13 @@ public interface UserDao {
 
     List<User> getAllUsers();
 
-    User getUser(int id);
+    User getUser(long id);
 
     void saveUser(User user);
 
-    void updateUser(int id, User user);
+    void updateUser(long id, User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     User findByEmail(String email);
 }
