+++ src/main/java/ru/kata/spring/boot_security/demo/services/UserService.java
@@ -8,13 +8,13 @@ public interface UserService {
 
     List<User> getAllUsers();
 
-    User getUser(int id);
+    User getUser(long id);
 
     void saveUser(User user);
 
-    void updateUser(int id, User user);
+    void updateUser(long id, User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     User findByEmail(String email);
 }
